Ask users to confirm deleting events w/ bookings
